+++ src/lib/components/Tabset/index.ts
@@ -1,2 +1,0 @@
-export * from './Tabset';
-export * from './Tab';